requirements.txt
torch==2.4.0
torchvision==0.20.0
torchaudio==2.4.0
numpy==1.26.4
matplotlib==3.9.0
pandas==2.2.2
tqdm==4.66.4
tensorboard==2.17.0
Pillow==10.4.0